refactor: rename Position component to SoptPart
