+++ src/views/IntroducePage/components/SoptPart/style.css.ts
