+++ src/views/IntroducePage/index.tsx
@@ -1,8 +1,8 @@
 import CoreValue from 'views/IntroducePage/components/CoreValue';
-import Position from 'views/IntroducePage/components/Position';
 import RecruitmentTarget from 'views/IntroducePage/components/RecruitmentTarget';
 import { wrapper } from './style.css';
 import { useDeviceType } from 'contexts/DeviceTypeProvider';
+import SoptPart from 'views/IntroducePage/components/SoptPart';
 
 const IntroducePage = () => {
   const { deviceType } = useDeviceType();
@@ -10,7 +10,7 @@ const IntroducePage = () => {
   return (
     <div className={wrapper[deviceType]}>
       <RecruitmentTarget />
-      <Position />
+      <SoptPart />
       <CoreValue />
     </div>
   );
